refactor netty4 and initialozaExtensionTransport method
