+++ src/test/java/org/opensearch/sdk/TestNetty4Transport.java
@@ -24,12 +24,12 @@ import org.opensearch.transport.netty4.Netty4Transport;
 
 public class TestNetty4Transport extends OpenSearchTestCase {
 
-    private ExtensionsRunner extensionsRunner;
     private ThreadPool threadPool;
+    private GetNetty4Transport getNetty4Transport = new GetNetty4Transport();
 
     @BeforeEach
     public void setUp() throws IOException {
-        this.extensionsRunner = new ExtensionsRunner();
+        new ExtensionsRunner();
         this.threadPool = new TestThreadPool("test");
     }
 
@@ -44,7 +44,7 @@ public class TestNetty4Transport extends OpenSearchTestCase {
             .put("transport.profiles.client1.port", 0)
             .build();
 
-        Netty4Transport transport = extensionsRunner.getNetty4Transport(settings, threadPool);
+        Netty4Transport transport = getNetty4Transport.getNetty4Transport(settings, threadPool);
 
         try {
             startNetty4Transport(transport);
@@ -67,7 +67,7 @@ public class TestNetty4Transport extends OpenSearchTestCase {
             .put("transport.profiles.client1.port", 0)
             .build();
 
-        Netty4Transport transport = extensionsRunner.getNetty4Transport(settings, threadPool);
+        Netty4Transport transport = getNetty4Transport.getNetty4Transport(settings, threadPool);
 
         try {
             startNetty4Transport(transport);
@@ -94,7 +94,7 @@ public class TestNetty4Transport extends OpenSearchTestCase {
             // attempt creating netty object with invalid settings
             IllegalStateException ex = expectThrows(
                 IllegalStateException.class,
-                () -> extensionsRunner.getNetty4Transport(settings, threadPool)
+                () -> getNetty4Transport.getNetty4Transport(settings, threadPool)
             );
             assertEquals("profile [no_port] has no port configured", ex.getMessage());
         } finally {
@@ -112,7 +112,7 @@ public class TestNetty4Transport extends OpenSearchTestCase {
             .put("transport.profiles.default.port", 0) // default port configuration will overwrite attempt
             .build();
 
-        Netty4Transport transport = extensionsRunner.getNetty4Transport(settings, threadPool);
+        Netty4Transport transport = getNetty4Transport.getNetty4Transport(settings, threadPool);
 
         try {
             startNetty4Transport(transport);
